+++ AmeisenBotX.Core/StateMachine/States/StateLoadingScreen.cs
@@ -27,7 +27,7 @@ namespace AmeisenBotX.Core.Statemachine.States
             if (WowInterface.XMemory.ReadString(WowInterface.OffsetList.GameState, Encoding.ASCII, out string gameState)
                 && gameState.Contains("login"))
             {
-                StateMachine.SetState((int)BotState.Login);
+                StateMachine.SetState(BotState.Login);
                 BotUtils.SendKey(WowInterface.XMemory.Process.MainWindowHandle, new IntPtr(0x1B));
                 return;
             }
@@ -35,7 +35,7 @@ namespace AmeisenBotX.Core.Statemachine.States
             WowInterface.ObjectManager.RefreshIsWorldLoaded();
             if (WowInterface.ObjectManager.IsWorldLoaded)
             {
-                StateMachine.SetState((int)BotState.Idle);
+                StateMachine.SetState(BotState.Idle);
                 return;
             }
         }
